change the save path of log file & upgrade deploy.sh

+++ deploy.sh
@@ -1,10 +1,19 @@
 #!/bin/sh
 
 PID=`ps -ef | grep "go-web-demo" | grep "server" | grep -v grep | awk '{print $2}'`
-kill $PID
-echo "The go-web-demo has stopped!"
+if [ -z $PID ]; then
+	echo "The server has yet to launch."
+else
+	kill $PID
+	echo "The server has stopped!"
+fi
 go build server
 chmod +rx server
 DIR=`pwd`
-$DIR/server 2>&1 | cronolog "go-web-demo.log.%Y-%m-%d" &
-echo "The go-web-demo has been launched."
+LOG_DIR=$DIR/logs
+if [ ! -d "$LOG_DIR" ]; then
+	echo "Creating dir '$LOG_DIR' ..."
+	mkdir "$LOG_DIR"
+fi
+$DIR/server 2>&1 | cronolog "$LOG_DIR/go-web-demo.log.%Y-%m-%d" &
+echo "The server has been launched."
